fix: Shared worker missing connect event
